Converter/ConverterAi/requirements.txt
aiohttp
aiobot
pillow
pydub
aspose-words
aiogram
Spire.Doc
Spire.Presentation
python-docx
pypandoc==1.5
pdflatex
pypandoc
pyPDF2
moviepy
yoomoney
